backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
openai==1.35.7
supabase==2.4.0
httpx==0.27.0 